feat: improve scripts

+++ home/.chezmoiscripts/unix/run_after_symlink-spaceship-theme.sh
@@ -3,16 +3,22 @@
 echo "zsh: Setting up Spaceship theme..."
 
 ZSH_CUSTOM="$HOME/.oh-my-zsh/custom"
+target_path="$ZSH_CUSTOM/themes/spaceship.zsh-theme"
 symlink_path="$ZSH_CUSTOM/themes/spaceship-prompt/spaceship.zsh-theme"
 
-# Check if the symbolic link doesn't exist
+# Check if the path exists
 if [[ ! -e "$symlink_path" ]]; then
-  target_path="$ZSH_CUSTOM/themes/spaceship.zsh-theme"
-
-  ln -s "$symlink_path" "$target_path"
-
-  echo "     Symbolic link for Spaceship theme created."
-  echo "     $symlink_path -> $target_path"
-else
-  echo "     Symbolic link for Spaceship theme already exists. Skipping."
+  echo "     Spaceship theme not found. Skipping."
+  return
 fi
+
+# Check if the symbolic link already exists
+if [[ -L "$target_path" ]]; then
+  echo "     Symbolic link for Spaceship theme already exists. Skipping."
+  return
+fi
+
+ln -s "$symlink_path" "$target_path"
+
+echo "     Symbolic link for Spaceship theme created."
+echo "     $symlink_path -> $target_path"
